fix: template root change
