+++ src/views/components/templates/post/Mdx/index.tsx
@@ -1,0 +1,75 @@
+/* eslint-disable react/no-danger */
+import React from 'react';
+import { graphql } from 'gatsby';
+
+import { MDXRenderer } from 'gatsby-plugin-mdx';
+import SEO from '@view/components/seo';
+import { rhythm, scale } from '@style/typography';
+import PostTemplate from '@template/post';
+
+interface Props {
+  data: {
+    mdx: any;
+    site: {
+      siteMetadata: {
+        title: string;
+      };
+    };
+  };
+  pageContext: any;
+}
+
+const MdxPostTemplate = ({ data }: Props) => {
+  const post = data.mdx;
+  return (
+    <PostTemplate
+      imgSrc={post.frontmatter.featuredImage?.childImageSharp.original.src}
+    >
+      <SEO
+        title={post.frontmatter.title}
+        description={post.frontmatter.description || post.excerpt}
+      />
+      <h1
+        style={{
+          marginBottom: 0,
+        }}
+      >
+        {post.frontmatter.title}
+      </h1>
+      <p
+        style={{
+          ...scale(-1 / 5),
+          display: 'block',
+          marginBottom: rhythm(1),
+        }}
+      >
+        {post.frontmatter.date}
+      </p>
+      <MDXRenderer>{post.body}</MDXRenderer>
+    </PostTemplate>
+  );
+};
+
+export default MdxPostTemplate;
+
+export const pageQuery = graphql`
+  query BlogPostBySlug($slug: String!) {
+    mdx(fields: { slug: { eq: $slug } }) {
+      id
+      excerpt(pruneLength: 160)
+      body
+      frontmatter {
+        title
+        date(formatString: "MMMM DD, YYYY")
+        description
+        featuredImage {
+          childImageSharp {
+            original {
+              src
+            }
+          }
+        }
+      }
+    }
+  }
+`;
